Move Contact and AppStatus modules into Settings
