+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/appstatus/AppStatusActivity.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.contact.appstatus
+package io.horizontalsystems.bankwallet.modules.settings.appstatus
 
 import android.os.Bundle
 import android.view.Menu
