+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/appstatus/AppStatusPresenter.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.contact.appstatus
+package io.horizontalsystems.bankwallet.modules.settings.appstatus
 
 import androidx.lifecycle.ViewModel
