+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/contact/ContactInteractor.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.contact
+package io.horizontalsystems.bankwallet.modules.settings.contact
 
 import io.horizontalsystems.bankwallet.core.IAppConfigProvider
 import io.horizontalsystems.bankwallet.core.IClipboardManager
