+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/contact/ContactModule.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.contact
+package io.horizontalsystems.bankwallet.modules.settings.contact
 
 import android.app.Activity
 import android.content.Intent
