+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/contact/ContactPresenter.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.contact
+package io.horizontalsystems.bankwallet.modules.settings.contact
 
 import androidx.lifecycle.ViewModel
