+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/contact/ContactRouter.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.contact
+package io.horizontalsystems.bankwallet.modules.settings.contact
 
 import io.horizontalsystems.core.SingleLiveEvent
